Refactors Get endpoint, moves functions to main module
* Get struct has a build function to contain all useful properties
