+++ lib/pocketeer/http_error.ex
@@ -1,0 +1,3 @@
+defmodule Pocketeer.HTTPError do
+  defexception [:message]
+end
