+++ lib/pocketeer/http_handler.ex
@@ -16,7 +16,7 @@ defmodule Pocketeer.HTTPHandler do
     case response do
       %HTTPotion.Response{body: body, headers: headers, status_code: status} when status in 200..299 ->
         {:ok, Response.new(status, headers, body) |> process_body}
-      %HTTPotion.Response{body: body, headers: headers, status_code: status} ->
+      %HTTPotion.Response{body: body, headers: headers, status_code: _status} ->
         {:error, %HTTPError{message: parse_error_message(body, headers)}}
       %HTTPotion.HTTPError{message: message} ->
         {:error, %HTTPError{message: message}}
